Rename schemas folder to models
