+++ app/api/api_v1/routers/admin.py
@@ -13,7 +13,7 @@ from sqlalchemy import Column, update
 
 from app.clients.db.session import get_db
 from app.repository.lookups import get_family_document_by_import_id_or_slug
-from app.schemas.document import DocumentUpdateRequest
+from app.models.document import DocumentUpdateRequest
 from app.service.auth import get_superuser_details
 
 _LOGGER = logging.getLogger(__name__)
