+++ app/api/api_v1/routers/geographies.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException, status
 from app.clients.db.session import get_db
 from app.errors import RepositoryError
 from app.repository.geography import get_world_map_stats
-from app.schemas.geography import GeographyStatsDTO
+from app.models.geography import GeographyStatsDTO
 
 _LOGGER = logging.getLogger(__file__)
