+++ app/api/api_v1/routers/pipeline_trigger.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import Session
 from app.clients.aws.client import S3Client, get_s3_client
 from app.clients.db.session import get_db
 from app.core.validation.util import get_new_s3_prefix, write_documents_to_s3
-from app.schemas.document import BulkIngestResult
+from app.models.document import BulkIngestResult
 from app.service.auth import get_superuser_details
 from app.service.pipeline import get_db_state_content
