+++ app/core/browse.py
@@ -11,7 +11,7 @@ from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
 from app.repository.geography import get_geo_subquery
-from app.schemas.search import (
+from app.models.search import (
     SearchResponse,
     SearchResponseFamily,
     SortField,
