+++ app/repository/pipeline.py
@@ -10,7 +10,7 @@ from db_client.models.dfce.family import FamilyDocument
 from fastapi import Depends
 
 from app.clients.db.session import get_db
-from app.schemas.document import DocumentParserInput
+from app.models.document import DocumentParserInput
 
 _LOGGER = logging.getLogger(__name__)
