+++ app/service/auth.py
@@ -6,7 +6,7 @@ from fastapi import Depends, HTTPException, status
 from jwt import PyJWTError
 
 from app.repository.user import get_app_user_by_email
-from app.schemas.user import JWTUser
+from app.models.user import JWTUser
 from app.service import security
 
 CREDENTIALS_EXCEPTION = HTTPException(
