+++ app/service/pipeline.py
@@ -5,7 +5,7 @@ from fastapi import Depends
 
 from app.clients.db.session import get_db
 from app.repository.pipeline import generate_pipeline_ingest_input
-from app.schemas.document import DocumentParserInput
+from app.models.document import DocumentParserInput
 
 _LOGGER = logging.getLogger(__name__)
