+++ tests/non_search/core/validation/test_util.py
@@ -8,7 +8,7 @@ import pytest
 from app.config import PIPELINE_BUCKET
 from app.core.validation import IMPORT_ID_MATCHER
 from app.core.validation.util import _flatten_maybe_tree, write_documents_to_s3
-from app.schemas.document import DocumentParserInput
+from app.models.document import DocumentParserInput
 
 NOT_A_TREE_1 = [{"name": 1}, {"name": 2}, {"name": 3}]
 NOT_A_TREE_2 = [
